BracketBuddyCron/requirements.txt
azure-core==1.23.1
azure-cosmos==4.2.0
graphqlclient==0.2.4
DateTime~=4.9
requests~=2.32.3
urllib3==2.3.0